Changed datatypes and added unit test

+++ src/Models/TransactionResults.cs
@@ -4,7 +4,7 @@
 	{
 		public string BlockHash { get; set; }
 
-		public int BlockNumber { get; set; }
+		public long BlockNumber { get; set; }
 
 		public string From { get; set; }
 
@@ -16,6 +16,6 @@
 
 		public decimal Value { get; set; }
 
-		public int TransactionIndex { get; set; }
+		public long TransactionIndex { get; set; }
 	}
 }
